FROMGIT: panel: simple: Fix size and bpp of BOE NV133FHM-N61
The BOE NV133FHM-N61 is documented in the original commit to be a
13.3" panel, but the size listed in our struct doesn't match.
Specifically:

  math.sqrt(30.0 * 30.0 + 18.7 * 18.7) / 2.54 ==> 13.92

Searching around on the Internet shows that the size that was in the
structure was the "Outline Size", not the "Display Area".  Let's fix
it.

Also the Internet says that this panel supports 262K colors.  That's
6bpp, not 8bpp.

Fixes: b0c664cc80e8 ("panel: simple: Add BOE NV133FHM-N61")
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Sam Ravnborg <sam@ravnborg.org>
Link: https://patchwork.freedesktop.org/patch/msgid/20200508155859.1.I4d29651c0837b4095fb4951253f44036a371732f@changeid
(cherry picked from commit 9694d9c3b5a5bbebd8132b08503211f5752dc205
 git://anongit.freedesktop.org/drm/drm-misc drm-misc-next)

BUG=b:153795921
TEST=Lazor panel works

Change-Id: I4d29651c0837b4095fb4951253f44036a371732f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2194043
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>