CHROMIUM: arm64: dts: qcom: Define that sc7180-trogdor has 2 eDP lanes
The sn65dsi86 can output up to 4 lanes of eDP.  We only have 2 lanes
connected on sc7180-trogdor boards.  Let's specify that.

This change is a no-op at the moment for a number of reasons:

* The driver already asks the panel for the max number of lanes and
  will never use more than the panel supports.  That means that this
  could only be a problem if someone hooked up a panel that supported
  4 lanes and then didn't change the board to route all 4 lanes.
  Panels aren't hot-swappable (we hardcode the panel name in the EDID)
  and we don't know of any cases where we've done this.

* The code currently doesn't parse this property (though it will
  soon).

Even though the change is an effective no-op, it is the correct thing
to do and it's specified in the bindings so let's do it now.

BUG=b:155686212
TEST=Trogdor and lazor screens still work

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ie8bde371625f9ef01847a4fab92f632368543633
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2192393
Reviewed-by: Stephen Boyd <swboyd@chromium.org>