FROMGIT: opp: Add sanity checks in _read_opp_key()
When we read the OPP keys, it would be nice to do some sanity checks
of the values we get from DT and see if they match with the information
that is populated in the OPP table. Let's pass a pointer of the table,
so that we can do some validation.

Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Sibi Sankar <sibis@codeaurora.org>
[ Viresh: Fix rebase conflicts ]
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
(cherry picked from commit 120e117bdc8483773fdae19dd326ef460b1ab4c2
 git://git.kernel.org/pub/scm/linux/kernel/git/vireshk/pm.git opp/linux-next)

BUG=b:111548643, b:149515210
TEST=build for trogdor

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I7381e1d02a28d20319b16759e29c6d4e44cc652b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2189602
Reviewed-by: Evan Green <evgreen@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>