CHROMIUM: iwl7000: pass number of tables in iwl_sar_select_profile() call
We were mistakenly passing the number of chains (which is 2) instead
of the number of tables (which is 1) to the iwl_sar_select_profile()
function.  Add a macro to define the number of tables and pass it
instead.

Change-Id: I36027aafe1ed049c9633ec8e244b4f6511195527
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: f23e426f2cd323ad88e4ad8ae0ca0a14ddd96ab3