UPSTREAM: drm/i915: Fix ivb cpu edp vswing
According to the DP spec supporting vswing 1 + preemph 2 is
mandatory. We don't have the hw settings for that though. In
order to pretend to follow the DP spec let's just select
vswing 0 + preemph 2 in this case (the DP spec says to use
the requested preemph in preference to the vswing when the
requested values aren't supported).

Signed-off-by: Ville Syrjälä <ville.syrjala@linux.intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20200512174145.3186-4-ville.syrjala@linux.intel.com
Reviewed-by: José Roberto de Souza <jose.souza@intel.com>
(cherry picked from commit 33520eae45f1c97670c8de7adec1cb5ab6a94416)



BUG=b:170142088
TEST=Compile tested/Backported fixes identified with r-o-m/On-device testing with eve-kernelnext

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I793e7867b81d04a44e865bb5bce884859da15ac2