UPSTREAM: power: supply: sbs-battery: add ability to disable charger broadcasts
In certain designs, it is possible to add a battery on a populated i2c
bus without an sbs compliant charger. In that case, the battery will
unnecessarily and sometimes undesirably master the bus trying to write
info in the charger.

It is observed in many occasion that these battery "broadcasts" are even
corrupting other ongoing master to slave communication. I.e. the
multi-master support in the battery is inadequate.

Thankfully, the CHARGER_MODE bit allows designers to disable that SBS
battery behaviour.

This needs to be done once when the battery is first seen on the bus.

Signed-off-by: Jean-Francois Dagenais <jeff.dagenais@gmail.com>
[rebased code]
Signed-off-by: Sebastian Reichel <sebastian.reichel@collabora.com>
(cherry picked from commit 182fc88268f3e33fb81cbb373b30d656daf0cb48)

BUG=b:161850135
TEST=read all sysfs knobs and check EC's console

Signed-off-by: Ikjoon Jang <ikjn@chromium.org>
Change-Id: Ie908af2c2b31f6e1702d77e62ed8ada1df2eadaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2465434
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>