CHROMIUM: iwl7000: mvm: avoid uninitialized memory access
When adding a PASN station, the station struct may be freed
before it was added to the linked list (if sending the dynamic
config command to FW fails). This may lead to uninitialized memory
access since the list pointers were not initialized yet.
Fix it by zeroing the station's struct memory on allocation to ensure
the list pointers are initialized to NULL.

Change-Id: I022ef88526cb3b95ab302593644d2e1b25bbd437
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
iwl7000-tree: 49444efc1cb05ffc05aedaf2a7ba9aea91662a09