Revert "FROMLIST: OPP: Add function to look up required OPP's for a given OPP"
The patch is part of the series "Add required-opps support to devfreq
passive gov" which was forward ported during the rebase from v4.9,
because it might eventually be needed by mt8192. With one patch of the
series NULL pointer dereferences in _set_required_opps() are observed
on trogdor with the "DVFS for IO devices on sdm845 and sc7180" series
(https://lore.kernel.org/patchwork/project/lkml/list/?series=443679).

Revert since it isn't clear whether the "Add required-opps support to
devfreq passive gov" series is actually needed. Also the last refresh
of the series was 10 months ago, if it is needed it should be in the
process of being upstreamed, as of now it seems abandoned.

BUG=b:80592702
TEST=boot trogdor with "DVFS for IO devices" series

This reverts commit d7248076cf2336eaf95971332982aeb0a62bd609.

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I39aa67f2ed94edcae22c8dda22496b1ed3fc61cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2202420
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>