Revert "FROMLIST: OPP: Allow required-opps even if the device doesn't have power-domains"
This reverts commit 2292d8d676959b353bf658da05b62c8fa2e919cb.

The patch is part of the series "Add required-opps support to devfreq
passive gov" which was forward ported during the rebase from v4.9,
because it might eventually be needed by mt8192. With one patch of the
series NULL pointer dereferences in _set_required_opps() are observed
on trogdor with the "DVFS for IO devices on sdm845 and sc7180" series
(https://lore.kernel.org/patchwork/project/lkml/list/?series=443679).

Revert since it isn't clear whether the "Add required-opps support to
devfreq passive gov" series is actually needed. Also the last refresh
of the series was 10 months ago, if it is needed it should be in the
process of being upstreamed, as of now it seems abandoned.

BUG=b:80592702
TEST=boot trogdor with "DVFS for IO devices" series

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I8818b99a748a392d4d1f630c617d86b8ece4d9ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2202421
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>