UPSTREAM: iommu/amd: Use pci_ats_supported()
The pci_ats_supported() function checks if a device supports ATS and is
allowed to use it. In addition to checking that the device has an ATS
capability and that the global pci=noats is not set
(pci_ats_disabled()), it also checks if a device is untrusted.

A device is untrusted if it is plugged into an external-facing port such
as Thunderbolt and could be spoofing an existing device to exploit
weaknesses in the IOMMU configuration. By calling pci_ats_supported() we
keep DTE[I]=0 for untrusted devices and abort transactions with
Pretranslated Addresses.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>
Reviewed-by: Joerg Roedel <jroedel@suse.de>
Link: https://lore.kernel.org/r/20200520152201.3309416-3-jean-philippe@linaro.org
Signed-off-by: Joerg Roedel <jroedel@suse.de>
(cherry picked from commit 7a441b2110527851f630144ec76ab8409e1d6c61)

BUG=b:155506690,b:152235212
TEST=build and boot volteer and verify external thunderbolt dock

Signed-off-by: Rajat Jain <rajatja@google.com>
Change-Id: I9a91851f4f8882786420aa93a58c6f337053c32c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2273950
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Alex Levin <levinale@google.com>
Reviewed-by: Benson Leung <bleung@google.com>