UPSTREAM: thunderbolt: Add trivial .shutdown
On my machine, a kexec with this driver loaded in the old kernel causes
a very long delay on boot in the kexec'ed kernel, most likely due to
unclean shutdown prior to that.

Unloading thunderbolt driver prior to kexec allows kexec to work as fast
as regular kernel boot, as well as adding this .shutdown pointer.

Shutting a device prior to the shutdown completely is always a good idea
IMHO to help with kexec, and this one-liner patch implements it.

BUG=None
TEST=Build and boot volteer; verify that SW CM and USB/PCIe tunnels
continue to work.

Signed-off-by: Maxim Levitsky <mlevitsk@redhat.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 4caf2511ec498277333d229f8a7ad1fa9331df65)
Signed-off-by: Prashant Malani <pmalani@chromium.org>
Change-Id: I87dfc947ab8a904c8a97f9050761716049b22c29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2258960
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>