FROMGIT: interconnect: Relax requirement in of_icc_get_from_provider()
This patch relaxes the condition in of_icc_get_from_provider() so that
it is no longer required to set '#interconnect-cells' to <1> in the DT,
and therefore it is not required to supply dummy node IDs in the
'interconnects' property when node IDs are dynamically generated rather
than hardcoded (statically allocated).

In case of the devfreq driver for exynos-bus, node IDs are dynamically
allocated and '#interconnect-cells' is always zero.

Signed-off-by: Artur Świgoń <a.swigon@samsung.com>
Acked-by: Krzysztof Kozlowski <krzk@kernel.org>
Reviewed-by: Chanwoo Choi <cw00.choi@samsung.com>
Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>
Link: https://lore.kernel.org/r/20200521122841.8867-3-s.nawrocki@samsung.com
Signed-off-by: Georgi Djakov <georgi.djakov@linaro.org>
(cherry picked from commit 0259a41da3233744f6a3e7848430da13100423d6
 https://git.linaro.org/people/georgi.djakov/linux.git icc-next)

BUG=b:149515210
TEST=build for trogdor

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: Ic8b82db931172c7165169c993bcb8a09d4afaf59
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2318352
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>