UPSTREAM: iio: adc: Add support for PMIC7 ADC
The ADC architecture on PMIC7 is changed as compared to PMIC5. The
major change from PMIC5 is that all SW communication to ADC goes through
PMK8350, which communicates with other PMICs through PBS when the ADC
on PMK8350 works in master mode. The SID register is used to identify the
PMICs with which the PBS needs to communicate. Add support for the same.

Signed-off-by: Jishnu Prakash <jprakash@codeaurora.org>
Reviewed-by: Andy Shevchenko <andy.shevchenko@gmail.com>
Signed-off-by: Jonathan Cameron <Jonathan.Cameron@huawei.com>
(cherry picked from commit 082111e559055de66da41303b15958c494c1b984)

BUG=b:165442834
TEST=pre-cq passes

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I077d38e9bf428c8045f6b79b9304c902fb9ad148
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2508994
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>