UPSTREAM: soundwire: intel: clarify drvdata and remove more indirections
The use of drvdata mixes two structures. There was no harm the first
structure is embedded as the first element of the second, but that's
not good. Make sure all drvdata is based on the 'sdw_cdns' structure.

While we are at it, remove indirections for 'dev' and 'cdns' to make
the code more readable.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>
Link: https://lore.kernel.org/r/20200531182102.27840-3-yung-chuan.liao@linux.intel.com
Signed-off-by: Vinod Koul <vkoul@kernel.org>
(cherry picked from commit 83e129afbe5c4a444c034e981b0f8535889380fe)
Signed-off-by: Chintan Patel <chintan.m.patel@intel.com>

BUG=b:162425092
TEST=On Soundwire SKU, audio playback via Speaker/Headphone/HDMI/DP

Change-Id: I24e8bb3018c4a835195ee8f9af78499146dbafed