CHROMIUM: arm64: dts: mt8192: asurada: enable necessary power
The below power rails need to keep always-on for the system
stability. We set them in the project.dtsi because there is no
kernel driver to handle them or the kernel driver cannot keep
them always-on due to the upstream common flow.
Please refer below infomation for the power:
Vgpu11 is for soc core power.
Vrf12 is for dram related power.
Vufs is for emmc power.

BUG=b:174105387
TEST=boot Asurada

Signed-off-by: Hsin-Hsiung Wang <hsin-hsiung.wang@mediatek.com>
Change-Id: Ic53e4527e173fd2d79a27d48a6679a1c1dea9e88
Signed-off-by: Yidi Lin <yidi.lin@mediatek.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2230078
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>
Commit-Queue: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>