FROMGIT: thunderbolt: Add runtime PM for Software CM
This adds runtime PM support for the Software Connection Manager parts
of the driver. This allows to save power when either there is no device
attached at all or there is a device attached and all following
conditions are true:

  - Tunneled PCIe root/downstream ports are runtime suspended
  - Tunneled USB3 ports are runtime suspended
  - No active DisplayPort stream
  - No active XDomain connection

For the first two we take advantage of device links that were added in
previous patch. Difference for the system sleep case is that we also
enable wakes when something is geting plugged in/out of the Thunderbolt
ports.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 6ac6faee5d7d7d1676a3188286438bed2dadd863
https://git.kernel.org/pub/scm/linux/kernel/git/westeri/thunderbolt.git next)

BUG=b:162700590
TEST=No regressions for the following use cases:
- S0iX NDA/DA
- TBT NVMe hotplug
- TBT dock hotplug
- TBT dock w/ USB pen drive
- TBT dock w/ display

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I242f962170847b8c7a8205d2caaebb6752240b75