FIXUP: FROMLIST: PM / OPP: Support adjusting OPP voltages at runtime
===
This is a squash of a revert of commit 4f56e9d413f8 ("FROMLIST:
PM / OPP: Support adjusting OPP voltages at runtime") AKA
CL:1628412 and a pick of what landed upstream.
===

On some SoCs the Adaptive Voltage Scaling (AVS) technique is
employed to optimize the operating voltage of a device. At a
given frequency, the hardware monitors dynamic factors and either
makes a suggestion for how much to adjust a voltage for the
current frequency, or it automatically adjusts the voltage
without software intervention. Add an API to the OPP library for
the former case, so that AVS type devices can update the voltages
for an OPP when the hardware determines the voltage should
change. The assumption is that drivers like CPUfreq or devfreq
will register for the OPP notifiers and adjust the voltage
according to suggestions that AVS makes.

This patch is derived from [1] submitted by Stephen.
[1] https://lore.kernel.org/patchwork/patch/599279/

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>
[Roger Lu: Changed to rcu less implementation]
Signed-off-by: Roger Lu <roger.lu@mediatek.com>
[s.nawrocki@samsung.com: added handling of OPP min/max voltage]
Signed-off-by: Sylwester Nawrocki <s.nawrocki@samsung.com>
Signed-off-by: Viresh Kumar <viresh.kumar@linaro.org>
(cherry picked from commit 25cb20a212a1f989385dfe23230817e69c62bee5)

BUG=b:129039370
TEST=boot to shell

Cq-Depend: chromium:2233356
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I70f03e1151513700eba7742d960231823a81ef47
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2233355
Reviewed-by: Evan Green <evgreen@chromium.org>