CHROMIUM: sched: Add /proc/pid/tasks/tid/latency_sensitive attribute
Upstream is moving to use uclamp which we would like. However, only 2 of
the 3 attributes are available in it (the clamp ones). The third
attribute - uclamp latency sensitive attribute is out-of-tree in Android
and available only for CPU controller CGroups. This patch tries to add
support for the third so that the 2 upstream ones can be used without
requiring CPU controller CGroups.

In Chrome OS, typically the latency sensitive attribute has been set
using schedtune CGroup. However the upstream uclamp interface uses CPU
controller CGroups. Even if we carry the out-of-tree patch to add the
latency sensitive attribute to uclamp-cgroups, we still cannot use it
for our usecase. The reason is currently, several threads in the
'foreground' CPU controller CGroup don't have the schedtune prefer-idle
(aka latency sensitive) attribute set. If we set this attribute for
'foreground', then all foreground tasks will get the attribute set,
which we don't want.

The solution is to add a temporary "latency sensitive" attribute to
procfs until upstream figures out what this interface should look like.
These patches aim to make the code selfcontained in their own
translation units so as to make the maintenance burden minimal.
For the other uclamp attributes (clamp min/max), those will be set via
the upstream sched_setattr(2).

BUG=b:147773166
TEST=boot

Change-Id: I9b872b081ab6e876c1751603c947e7d9a7c210d9
Signed-off-by: Joel Fernandes <joelaf@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2241891
Reviewed-by: Brian Geffon <bgeffon@chromium.org>