CHROMIUM: media: camx: soc: Fix-up drivers probe order
Because of the new dependencies with Interconnect, Genpd and
Clock FW's the order of drivers probing can be changed
unpredictably. The reference camera code isn't prepared to
handle that situation.

The hierarchy of DTS was changed to perform to proper
sequence of drivers probes. The drivers are updated to
handle probe deferring.

BUG=b:168139312
TEST=Build

Signed-off-by: Atanas Filipov <afilipov@mm-sol.com>
Change-Id: Icf2a772cad7aeff20a4e790bad681540f7de9546
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2323420
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Commit-Queue: Tomasz Figa <tfiga@chromium.org>