Revert "UPSTREAM: Input: elants_i2c - report resolution information for touch major"
This reverts commit 6d3ced00d7e603fca970745b343b26995bbbf895.

This commit is causing b/158477364 on Dru and part of b/159195016 on
Waddledoo, so it seems that there are multiple ELAN touchscreens for
which the resolution of 1 unit per mm is incorrect. Reverting on all
Kernel versions seems like a prudent step until we have a more universal
way to set the resolution.

BUG=b:158477364,b:159195016
TEST=none

Signed-off-by: Harry Cutts <hcutts@chromium.org>
Change-Id: Idba365d226ee511f69aaffc456031554ce0bbdfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2252783
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>