UPSTREAM: drm/msm/dpu: move some sspp caps to dpu_caps
This isn't something that ever changes between planes, so move it to
dpu_caps struct. Making this change will allow more re-use in the
"SSPP sub blocks config" part of the catalog, in particular when adding
support for SM8150 and SM8250 which have different max_linewidth.

This also sets max_hdeci_exp/max_vdeci_exp to 0 for sc7180, as decimation
is not supported on the newest DPU versions. (note that decimation is not
implemented, so this changes nothing)

Signed-off-by: Jonathan Marek <jonathan@marek.ca>
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 7e9d4cdd653a4e9de6474fbca05c095f15798188)

BUG=b:141308430
TEST=boot lazor, test a few games, etc

Signed-off-by: Rob Clark <robdclark@chromium.org>
Change-Id: I94a30152dde7f5998419ab5252428776ea0af6ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2360850
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>