UPSTREAM: drm/msm: handle for EPROBE_DEFER for of_icc_get
Check for errors instead of silently not using icc if the msm driver
probes before the interconnect driver.

Allow ENODATA for ocmem path, as it is optional and this error
is returned when "gfx-mem" path is provided but not "ocmem".

Because msm_gpu_cleanup assumes msm_gpu_init has been called, the icc path
init needs to be after msm_gpu_init for the error path to work.

Signed-off-by: Jonathan Marek <jonathan@marek.ca>
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 08af4769c7d211ff800a143b60855ad7146befc8)

BUG=b:141308430
TEST=boot lazor, test a few games, etc

Signed-off-by: Rob Clark <robdclark@chromium.org>
Change-Id: I9e37693fd28beed63645f4be2b141c4c96d01da9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2360836
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>