FROMGIT: coresight: replicator: Reset replicator if context is lost
On some QCOM SoCs, replicators in Always-On domain loses its
context as soon as the clock is disabled. Currently as a part
of pm_runtime workqueue, clock is disabled after the replicator
is initialized by amba_pm_runtime_suspend assuming that context
is not lost which is not true for replicators with such
limitations. So add a new property "qcom,replicator-loses-context"
to identify such replicators and reset them.

Suggested-by: Mike Leach <mike.leach@linaro.org>
Reviewed-by: Mike Leach <mike.leach@linaro.org>
Signed-off-by: Sai Prakash Ranjan <saiprakash.ranjan@codeaurora.org>
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Link: https://lore.kernel.org/r/20200716175746.3338735-6-mathieu.poirier@linaro.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 8f3ce74c20f21e09644d2a584238f1453a00eb8d
 https://git.kernel.org/pub/scm/linux/kernel/git/gregkh/char-misc.git char-misc-next)

BUG=b:149851136
TEST=See /sys/bus/coresight has devices on Trogdor

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I6234f56ca53c74c7c5e4327c9d8e63f246fa6047
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2210952
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>