UPSTREAM: mei: hdcp: Replace one-element array with flexible-array member
There is a regular need in the kernel to provide a way to declare
having a dynamically sized set of trailing elements in a structure.
Kernel code should always use “flexible array members”[1] for these
cases. The older style of one-element or zero-length arrays should
no longer be used[2].

Also, make use of the array_size() helper instead of the open-coded
version in memcpy(). These sorts of multiplication factors need to
be wrapped in array_size().

And while there, use the preferred form for passing a size of a struct.
The alternative form where struct name is spelled out hurts readability
and introduces an opportunity for a bug when the pointer variable type is
changed but the corresponding sizeof that is passed as argument is not.

[1] https://en.wikipedia.org/wiki/Flexible_array_member
[2] https://github.com/KSPP/linux/issues/79

Signed-off-by: Gustavo A. R. Silva <gustavoars@kernel.org>
Link: https://lore.kernel.org/r/20200722181534.GA31357@embeddedor
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit c56967d674e361ebe716e66992e3c5332b25ac1f)

BUG=b:155511178
TEST=None

Signed-off-by: Sean Paul <seanpaul@chromium.org>
Change-Id: I5bb02c4f6723485fd2e38d66272286d5aff3395c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2587304
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>